+++ omf-core-framework/src/main/java/com/samares_engineering/omf/omf_core_framework/utils/utils/UMLUtils.kt
@@ -3,73 +3,104 @@
  * @Licence: EPL 2.0
  * @Author:   Quentin Cespédès, Clément Mezerette, Hugo Stinson
  * @since     0.0.0
- ******************************************************************************/
+ */
+package com.samares_engineering.omf.omf_core_framework.utils.utils
 
-package com.samares_engineering.omf.omf_core_framework.utils.utils;
+import com.nomagic.uml2.ext.jmi.helpers.StereotypesHelper
+import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element
+import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.NamedElement
+import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Property
+import com.nomagic.uml2.ext.magicdraw.mdprofiles.Stereotype
+import com.samares_engineering.omf.omf_core_framework.errors.exceptions.GenericException
+import com.samares_engineering.omf.omf_core_framework.errors.exceptions.core.LegacyOMFException
+import com.samares_engineering.omf.omf_core_framework.utils.OMFUtils
+import com.samares_engineering.omf.omf_core_framework.utils.utils.UMLUtils.getAllDerivedProperties
+import java.util.*
+import java.util.function.Function
+import java.util.stream.Collectors
 
-import com.nomagic.uml2.ext.jmi.helpers.StereotypesHelper;
-import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Class;
-import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Element;
-import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.NamedElement;
-import com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Property;
-import com.nomagic.uml2.ext.magicdraw.mdprofiles.Stereotype;
-import com.samares_engineering.omf.omf_core_framework.errors.exceptions.GenericException;
-import com.samares_engineering.omf.omf_core_framework.errors.exceptions.core.LegacyOMFException;
-import com.samares_engineering.omf.omf_core_framework.utils.OMFUtils;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-public class UMLUtils {
-    private UMLUtils() {}
-
-    public static List<Class> getStereotypeMetaClass(Stereotype str) {
-        return StereotypesHelper.getBaseClasses(str);
+object UMLUtils {
+    @JvmStatic
+    fun getStereotypeMetaClass(str: Stereotype?): List<com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Class> {
+        return StereotypesHelper.getBaseClasses(str)
     }
 
-    public static boolean isStereotypeMetaClassMatchClass(Stereotype str, java.lang.Class clazz) {
+    fun isStereotypeMetaClassMatchClass(str: Stereotype?, clazz: Class<*>?): Boolean {
         return getStereotypeMetaClass(str).stream()
-                .anyMatch(metaClass -> metaClass == StereotypesHelper.getMetaClassByClass(OMFUtils.getProject(), clazz));
+            .anyMatch { metaClass: com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Class ->
+                metaClass === StereotypesHelper.getMetaClassByClass(
+                    OMFUtils.getProject(),
+                    clazz
+                )
+            }
     }
 
-    public static boolean isInstanceOfMetaClass(Element type, java.lang.Class metaClass) throws LegacyOMFException {
+    @JvmStatic
+    @Throws(LegacyOMFException::class)
+    fun isInstanceOfMetaClass(type: Element?, metaClass: Class<*>): Boolean {
         try {
-            if (type == null)
-                return false;
+            if (type == null) return false
 
-            if (type instanceof Stereotype)
-                return isStereotypeMetaClassMatchClass((Stereotype) type, metaClass);
+            if (type is Stereotype) return isStereotypeMetaClassMatchClass(type as Stereotype?, metaClass)
 
-            return type instanceof NamedElement && ((NamedElement) type).getName().equals(metaClass.getSimpleName());
-        } catch (Exception e) {
-            throw new LegacyOMFException("Impossible to determine MetaClass of selected type",
-                    e, GenericException.ECriticality.CRITICAL);
+            return type is NamedElement && type.name == metaClass.simpleName
+        } catch (e: Exception) {
+            throw LegacyOMFException(
+                "Impossible to determine MetaClass of selected type",
+                e, GenericException.ECriticality.CRITICAL
+            )
         }
     }
 
-    public static Set<Property> getAllDerivedProperties(Stereotype stereotype) {
-        return StereotypesHelper.getPropertiesWithDerived(stereotype);
+    @JvmStatic
+    fun getAllMetaClasses(stereotypes: Collection<Stereotype>): List<Class<*>> {
+        return  stereotypes.asSequence()
+            .map { str: Stereotype -> UMLUtils.getStereotypeMetaClass(str) }
+            .flatten()
+            .distinct()
+            .filter(Objects::nonNull)
+            .map { mdClass: com.nomagic.uml2.ext.magicdraw.classes.mdkernel.Class ->
+                StereotypesHelper.getClassOfMetaClass(
+                    mdClass
+                )
+            }
+            .toList()
     }
 
-    public static Set<Property> getAllDerivedProperties(Element element) {
-        return element.getAppliedStereotype().stream()
-                .map(UMLUtils::getAllDerivedProperties)
-                .flatMap(Set::stream)
-                .collect(Collectors.toSet());
+    @JvmStatic
+    fun getAlMetaClassesAsArray(stereotypes: Collection<Stereotype>): Array<Class<*>> {
+        return getAllMetaClasses(stereotypes).toTypedArray()
     }
 
-    public static Optional<Property> getDerivedPropertyByName(Stereotype stereotype, String propertyName) {
-        return getAllDerivedProperties(stereotype).stream()
-                .filter(property -> property.getName().equals(propertyName))
-                .findAny();
+
+
+    /* ********************************************************************************************************************
+     *  Get all derived properties of a stereotype
+     * ********************************************************************************************************************/
+
+    @JvmStatic
+    fun getAllDerivedProperties(stereotype: Stereotype): Set<Property> {
+        return StereotypesHelper.getPropertiesWithDerived(stereotype)
     }
 
-    public static Optional<Property> getDerivedPropertyByName(Element element, String propertyName) {
-        return getAllDerivedProperties(element).stream()
-                .filter(property -> property.getName().equals(propertyName))
-                .findAny();
+    @JvmStatic
+    fun getAllDerivedProperties(element: Element): Set<Property> {
+        return element.appliedStereotype
+            .map{getAllDerivedProperties(it)}
+            .flatten()
+            .toSet()
     }
 
+    @JvmStatic
+    fun getDerivedPropertyByName(stereotype: Stereotype, propertyName: String): Property? {
+        return getAllDerivedProperties(stereotype)
+            .find { property: Property -> property.name == propertyName }
+
+    }
+
+    @JvmStatic
+    fun getDerivedPropertyByName(element: Element, propertyName: String): Property? {
+        return getAllDerivedProperties(element)
+            .find { property: Property -> property.name == propertyName }
+    }
 }